pyproject.toml
[tool.poetry]
name = "tweet-to-icon-gan"
version = "0.1.0"
description = ""
authors = ["kishimoto-banana <drehbleistift@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tweepy = "^3.8.0"
mecab-python3 = "^0.996.5"
mojimoji = "^0.0.10"
gensim = "^3.8.3"
torch = "^1.5.0"
torchvision = "^0.6.0"
Pillow = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
